backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
email-validator==2.2.0
pydantic-settings==2.7.0
motor==3.6.0
python-dotenv==1.0.1
python-dateutil==2.9.0
neo4j==5.27.0
univers==30.12.1
GitPython==3.1.43
toml==0.10.2
regex==2024.11.6
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pyjwt==2.10.1
aiohttp==3.11.11
aiocache==0.12.3
bcrypt==3.2.0
aiofiles==24.1.0